move cogito ProdPutter to its own file
